icd_code_extraction_using_gpt_RAG/insuranceLLM-master/requirements.txt
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
apache-beam==2.50.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==2.2.1
contourpy==1.1.1
crcmod==1.7
cycler==0.12.0
dataclasses-json==0.5.14
datasets==2.14.5
dill==0.3.1.1
dnspython==2.4.2
docopt==0.6.2
exceptiongroup==1.1.3
fastapi==0.103.2
fastavro==1.8.3
fasteners==0.19
ffmpy==0.3.1
filelock==3.12.4
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.6.0
googleapis-common-protos==1.60.0
gradio==3.45.2
gradio_client==0.5.3
greenlet==3.0.0
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.59.0
h11==0.14.0
hdfs==2.7.2
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
huggingface-hub==0.17.3
idna==3.4
importlib-resources==6.1.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.162
loguru==0.7.2
lz4==4.3.2
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
multidict==6.0.4
multiprocess==0.70.9
mwparserfromhell==0.6.5
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.24.4
objsize==0.6.1
openai==0.28.1
openapi-schema-pydantic==1.2.4
orjson==3.9.7
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
pinecone-client==2.2.4
proto-plus==1.22.3
protobuf==3.20.3
pyarrow==11.0.0
pydantic==1.10.13
pydantic_core==2.10.1
pydot==1.4.2
pydub==0.25.1
pymongo==4.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.10.3
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.5.1
toolz==0.12.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
websockets==11.0.3
xxhash==3.3.0
yarl==1.9.2
zstandard==0.21.0
flask==3.0.0
streamlit==1.27.2